LayoutTests/imported/w3c/web-platform-tests/cookie-store/serviceworker_cookieStore_cross_origin.https.sub-expected.txt

Blocked access to external URL https://www1.localhost:9443/cookie-store/resources/helper_iframe.sub.html
CONSOLE MESSAGE: Unable to post message to https://www1.localhost:9443. Recipient has origin https://localhost:9443.


Harness Error (TIMEOUT), message = null

TIMEOUT cookieStore.get() in ServiceWorker reads cookie set in cross-origin frame Test timed out
